requirements.txt
# 배민외식업광장 모니터링 - 로컬 PC용

# Selenium
selenium>=4.15.0

# ChromeDriver 자동 관리
webdriver-manager>=4.0.1

# Google Sheets API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# HTTP 요청
requests>=2.31.0
